+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - New function to load user's mentions timeline (`mentions_timeline`)
 - New function to load the posts by a given user (`user_timeline`)
 - New function to load the user's posts that have been retweeted (`retweets_of_me`)
+- New function to load the user IDs who have retweeted a given tweet (`retweeters_of`)
 
 ### Changed
 - Moved `UserID` into the user module